+++ app.js
@@ -1,54 +1,19 @@
 
 /**
 	* Node.js Login Boilerplate
-	* More Info : https://github.com/braitsch/node-login
-	* Copyright (c) 2013-2020 Stephen Braitsch
+	* https://node-login.braitsch.io
+	* (c) 2013-2021 Stephen Braitsch
 **/
 
-var http = require('http');
-var express = require('express');
-var session = require('express-session');
-var bodyParser = require('body-parser');
-var cookieParser = require('cookie-parser');
-var MongoStore = require('connect-mongo')(session);
+//	https://www.npmjs.com/package/@braitsch/express
+const express = require('@braitsch/express');
 
-var app = express();
+const app = express();
 
-app.locals.pretty = true;
-app.set('port', process.env.PORT || 3000);
-app.set('views', __dirname + '/app/server/views');
-app.set('view engine', 'pug');
-app.use(cookieParser());
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(require('stylus').middleware({ src: __dirname + '/app/public' }));
-app.use(express.static(__dirname + '/app/public'));
+express.log('./logs');
 
-// build mongo database connection url //
+express.http(app);
 
-process.env.DB_HOST = process.env.DB_HOST || 'localhost'
-process.env.DB_PORT = process.env.DB_PORT || 27017;
-process.env.DB_NAME = process.env.DB_NAME || 'node-login';
+express.init(__dirname, app, 'node-login', true);
 
-if (app.get('env') != 'live'){
-	process.env.DB_URL = 'mongodb://'+process.env.DB_HOST+':'+process.env.DB_PORT;
-}	else {
-// prepend url with authentication credentials // 
-	process.env.DB_URL = 'mongodb://'+process.env.DB_USER+':'+process.env.DB_PASS+'@'+process.env.DB_HOST+':'+process.env.DB_PORT;
-}
-
-app.use(session({
-	secret: 'faeb4453e5d14fe6f6d04637f78077c76c73d1b4',
-	proxy: true,
-	resave: true,
-	saveUninitialized: true,
-	store: new MongoStore({ url: process.env.DB_URL })
-	})
-);
-
-require('./app/server/routes')(app);
-
-http.createServer(app).listen(app.get('port'), function(){
-	console.log('Express server listening on port ' + app.get('port'));
-});
-
+express.start(app);