+++ server/routes.js
@@ -1,7 +1,7 @@
 
-var CT = require('./modules/country-list');
-var AM = require('./modules/account-manager');
-var EM = require('./modules/email-dispatcher');
+const accounts = require('./model/accounts');
+const emailjs = require('./utils/emailjs');
+const countries = require('./json/countries');
 
 module.exports = function(app) {
 
@@ -15,9 +15,9 @@ module.exports = function(app) {
 			res.render('login', { title: 'Hello - Please Login To Your Account' });
 		}	else{
 	// attempt automatic login //
-			AM.validateLoginKey(req.cookies.login, req.ip, function(e, o){
+			accounts.validateLoginKey(req.cookies.login, req.ip, function(e, o){
 				if (o){
-					AM.autoLogin(o.user, o.pass, function(o){
+					accounts.autoLogin(o.user, o.pass, function(o){
 						req.session.user = o;
 						res.redirect('/home');
 					});
@@ -27,9 +27,9 @@ module.exports = function(app) {
 			});
 		}
 	});
-	
+
 	app.post('/', function(req, res){
-		AM.manualLogin(req.body['user'], req.body['pass'], function(e, o){
+		accounts.manualLogin(req.body['user'], req.body['pass'], function(e, o){
 			if (!o){
 				res.status(400).send(e);
 			}	else{
@@ -37,7 +37,7 @@ module.exports = function(app) {
 				if (req.body['remember-me'] == 'false'){
 					res.status(200).send(o);
 				}	else{
-					AM.generateLoginKey(o.user, req.ip, function(key){
+					accounts.generateLoginKey(o.user, req.ip, function(key){
 						res.cookie('login', key, { maxAge: 900000 });
 						res.status(200).send(o);
 					});
@@ -50,28 +50,28 @@ module.exports = function(app) {
 		res.clearCookie('login');
 		req.session.destroy(function(e){ res.status(200).send('ok'); });
 	})
-	
+
 /*
 	control panel
 */
-	
+
 	app.get('/home', function(req, res) {
 		if (req.session.user == null){
 			res.redirect('/');
 		}	else{
 			res.render('home', {
 				title : 'Control Panel',
-				countries : CT,
+				countries : countries,
 				udata : req.session.user
 			});
 		}
 	});
-	
+
 	app.post('/home', function(req, res){
 		if (req.session.user == null){
 			res.redirect('/');
 		}	else{
-			AM.updateAccount({
+			accounts.updateAccount({
 				id		: req.session.user._id,
 				name	: req.body['name'],
 				email	: req.body['email'],
@@ -93,11 +93,11 @@ module.exports = function(app) {
 */
 
 	app.get('/signup', function(req, res) {
-		res.render('signup', {  title: 'Signup', countries : CT });
+		res.render('signup', {  title: 'Signup', countries : countries });
 	});
-	
+
 	app.post('/signup', function(req, res){
-		AM.addNewAccount({
+		accounts.addNewAccount({
 			name 	: req.body['name'],
 			email 	: req.body['email'],
 			user 	: req.body['user'],
@@ -118,17 +118,17 @@ module.exports = function(app) {
 
 	app.post('/lost-password', function(req, res){
 		let email = req.body['email'];
-		AM.generatePasswordKey(email, req.ip, function(e, account){
+		accounts.generatePasswordKey(email, req.ip, function(e, account){
 			if (e){
-				res.status(400).send(e);
+				res.status(404).send(e);
 			}	else{
-				EM.dispatchResetPasswordLink(account, function(e, m){
+				emailjs.dispatchResetPasswordLink(account, function(e){
 			// TODO this callback takes a moment to return, add a loader to give user feedback //
 					if (!e){
 						res.status(200).send('ok');
 					}	else{
-						for (k in e) console.log('ERROR : ', k, e[k]);
-						res.status(400).send('unable to dispatch password reset');
+						log(e);
+						res.status(500).send('unable to dispatch password reset');
 					}
 				});
 			}
@@ -136,7 +136,7 @@ module.exports = function(app) {
 	});
 
 	app.get('/reset-password', function(req, res) {
-		AM.validatePasswordKey(req.query['key'], req.ip, function(e, o){
+		accounts.validatePasswordKey(req.query['key'], req.ip, function(e, o){
 			if (e || o == null){
 				res.redirect('/');
 			} else{
@@ -145,13 +145,13 @@ module.exports = function(app) {
 			}
 		})
 	});
-	
+
 	app.post('/reset-password', function(req, res) {
 		let newPass = req.body['pass'];
 		let passKey = req.session.passKey;
 	// destory the session immediately after retrieving the stored passkey //
 		req.session.destroy();
-		AM.updatePassword(passKey, newPass, function(e, o){
+		accounts.updatePassword(passKey, newPass, function(e, o){
 			if (o){
 				res.status(200).send('ok');
 			}	else{
@@ -159,19 +159,19 @@ module.exports = function(app) {
 			}
 		})
 	});
-	
+
 /*
 	view, delete & reset accounts
 */
-	
+
 	app.get('/print', function(req, res) {
-		AM.getAllRecords( function(e, accounts){
+		accounts.getAllRecords( function(e, accounts){
 			res.render('print', { title : 'Account List', accts : accounts });
 		})
 	});
-	
+
 	app.post('/delete', function(req, res){
-		AM.deleteAccount(req.session.user._id, function(e, obj){
+		accounts.deleteAccount(req.session.user._id, function(e, obj){
 			if (!e){
 				res.clearCookie('login');
 				req.session.destroy(function(e){ res.status(200).send('ok'); });
@@ -180,13 +180,13 @@ module.exports = function(app) {
 			}
 		});
 	});
-	
+
 	app.get('/reset', function(req, res) {
-		AM.deleteAllAccounts(function(){
+		accounts.deleteAllAccounts(function(){
 			res.redirect('/print');
 		});
 	});
-	
+
 	app.get('*', function(req, res) { res.render('404', { title: 'Page Not Found'}); });
 
 };